+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## UNRELEASE
 - **Fix** : Alma issues on Woocommerce
-
+- **Fix** : Fixed shipping address for Paypal v2
 
 ## 2.9.1